Move unicodePrintable and jsonPrintable to it's own file, and add unit test
